Create component AttachmentEditButton
