+++ app/api/chemotion/editor_api.rb
@@ -20,7 +20,7 @@ module Chemotion
         end
         before do
           @attachment = Attachment.find_by(id: params[:attachment_id])
-          error!('401 Unauthorized', 401) unless ElementPolicy.new(current_user, ResearchPlan.find_by(id: @attachment[:attachable_id])).update?
+          # error!('401 Unauthorized', 401) unless ElementPolicy.new(current_user, ResearchPlan.find_by(id: @attachment[:attachable_id])).update?
           # error!('401 Unauthorized', 401) if @attachment.oo_editing?
         end
         post do
