+++ app/packs/src/components/generic/GenericAttachments.js
@@ -2,13 +2,11 @@ import React, { Component } from 'react';
 import PropTypes from 'prop-types';
 import Dropzone from 'react-dropzone';
 import { FormGroup, Button, Row, Col, Tooltip, ControlLabel, ListGroup, ListGroupItem, OverlayTrigger } from 'react-bootstrap';
-import { last, findKey, values } from 'lodash';
-import EditorFetcher from 'src/fetchers/EditorFetcher';
+import { values } from 'lodash';
 import ImageModal from 'src/components/common/ImageModal';
-import SpinnerPencilIcon from 'src/components/common/SpinnerPencilIcon';
 import { previewAttachmentImage } from 'src/utilities/imageHelper';
 import Utils from 'src/utilities/Functions';
-import NotificationActions from 'src/stores/alt/actions/NotificationActions';
+import AttachmentEditButton from 'src/apps/mydb/elements/details/AttachmentEditButton'
 
 const editorTooltip = exts => <Tooltip id="editor_tooltip">Available extensions: {exts}</Tooltip>;
 const downloadTooltip = <Tooltip id="download_tooltip">Download attachment</Tooltip>;
@@ -17,47 +15,7 @@ const defaultImgStyle = { position: 'absolute', width: 60, height: 60 };
 export default class GenericAttachments extends Component {
   constructor(props) {
     super(props);
-    this.state = { attachmentEditor: false, extension: null };
-    this.editorInitial = this.editorInitial.bind(this);
-  }
-
-  componentDidMount() {
-    this.editorInitial();
-  }
-
-  editorInitial() {
-    EditorFetcher.initial()
-      .then((result) => {
-        this.setState({ attachmentEditor: result.installed, extension: result.ext });
-      });
-  }
-
-  documentType(filename) {
-    const { extension } = this.state;
-    const ext = last(filename.split('.'));
-    const docType = findKey(extension, o => o.includes(ext));
-    if (typeof (docType) === 'undefined' || !docType) { return null; }
-    return docType;
-  }
-
-  handleEdit(att) {
-    const { onEdit } = this.props;
-    const attachment = att;
-    const fileType = last(attachment.filename.split('.'));
-    const docType = this.documentType(attachment.filename);
-    EditorFetcher.startEditing({ attachment_id: attachment.id })
-      .then((result) => {
-        if (result.token) {
-          const url = `/editor?id=${attachment.id}&docType=${docType}&fileType=${fileType}&title=${attachment.filename}&key=${result.token}`;
-          window.open(url, '_blank');
-          attachment.aasm_state = 'oo_editing';
-          attachment.updated_at = new Date();
-          onEdit(attachment);
-        } else {
-          // alert('Unauthorized to edit this file.');
-          NotificationActions.add({ message: 'Unauthorized to edit this file.', level: 'error', position: 'tc' });
-        }
-      });
+    this.state = { extension: null };
   }
 
   renderRemoveAttachmentButton(attachment) {
@@ -70,10 +28,8 @@ export default class GenericAttachments extends Component {
   }
 
   renderListGroupItem(attachment) {
-    const { attachmentEditor, extension } = this.state;
+    const { extension } = this.state;
     const { onDelete } = this.props;
-    const updateTime = new Date(attachment.updated_at);
-    updateTime.setTime(updateTime.getTime() + (15 * 60 * 1000));
     const fetchId = attachment.id;
     const previewImg = previewAttachmentImage(attachment);
     let imgStyle = { ...defaultImgStyle, cursor: 'pointer' };
@@ -84,10 +40,6 @@ export default class GenericAttachments extends Component {
       hasPop = false;
       fetchNeeded = false;
     }
-    const isEditing = attachment.aasm_state === 'oo_editing' && new Date().getTime() < updateTime;
-    const docType = this.documentType(attachment.filename);
-    const editDisable = !attachmentEditor || isEditing || attachment.is_new || docType === null;
-    const styleEditor = !attachmentEditor || docType === null ? 'none' : '';
     if (attachment.is_deleted) {
       return (
         <div>
@@ -132,11 +84,12 @@ export default class GenericAttachments extends Component {
                 <i className="fa fa-download" aria-hidden="true" />
               </Button>
             </OverlayTrigger>
-            <OverlayTrigger placement="left" overlay={editorTooltip(values(extension).join(','))} >
-              <Button style={{ display: styleEditor }} bsSize="xsmall" className="button-right" bsStyle="success" disabled={editDisable} onClick={() => this.handleEdit(attachment)}>
-                <SpinnerPencilIcon spinningLock={!attachmentEditor || isEditing} />
-              </Button>
-            </OverlayTrigger>
+            <AttachmentEditButton
+              attachment={attachment}
+              className="button-right"
+              overlay={editorTooltip(values(extension).join(','))}
+              onEdit={this.props.onEdit}
+            ></AttachmentEditButton>
           </Col>
         </Row>
       </div>
