+++ app/views/pages/editor.html.erb
@@ -59,7 +59,7 @@
                     documentType: docType,
                     document: {
                         title: title,
-                        url: callback_server + '/api/v1/public/download?token=' + key ,
+                        url: callback_server + '/api/v1/public/download?token=' + key,
                         fileType: "<%= params[:fileType] %>",
                         key: "<%= params[:key] %>",
                         info: {
